+++ src/rsdict/constants.rs
@@ -1,0 +1,4 @@
+pub const SMALL_BLOCK_SIZE: u64  = 64;
+pub const LARGE_BLOCK_SIZE: u64  = 1024;
+pub const SELECT_BLOCK_SIZE: u64 = 4096;
+pub const SMALL_BLOCK_PER_LARGE_BLOCK: u64 = LARGE_BLOCK_SIZE / SMALL_BLOCK_SIZE;
